requirements.txt
Flask
Requests
opentelemetry-api==0.14b0
opentelemetry-sdk==0.14b0
opentelemetry-ext-jaeger==0.6b0
opentelemetry-ext-flask==0.6b0
opentelemetry-ext-requests==0.11b0
opentelemetry-auto-instrumentation==0.8b0
opentelemetry-ext-lightstep==0.6b0
opentelemetry-ext-honeycomb==0.14b0